Move apply_type() to applytype.py
